requirements.txt
streamlit
yt-dlp
assemblyai
fpdf
PyPDF2
python-dotenv
requests
streamlit-extras
